Extract ScreenMenu component out of title screen.
